Adjust migrations
